Rename API folder
