+++ src/main/java/pl/plajer/pinata/api/PinataDeathEvent.java
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package pl.plajer.pinata.pinataapi;
+package pl.plajer.pinata.api;
 
 import java.util.List;
