+++ src/main/java/pl/plajer/pinata/commands/GameCommands.java
@@ -45,7 +45,7 @@ import pl.plajer.pinata.Main;
 import pl.plajer.pinata.handlers.language.LanguageManager;
 import pl.plajer.pinata.pinata.Pinata;
 import pl.plajer.pinata.pinata.PinataItem;
-import pl.plajer.pinata.pinataapi.PinataFactory;
+import pl.plajer.pinata.api.PinataFactory;
 import pl.plajer.pinata.utils.PinataUtils;
 import pl.plajer.pinata.utils.Utils;
 import pl.plajerlair.core.utils.ConfigUtils;
