+++ src/main/java/pl/plajer/pinata/handlers/MenuHandler.java
@@ -21,7 +21,6 @@ package pl.plajer.pinata.handlers;
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
 import org.bukkit.Location;
-import org.bukkit.Material;
 import org.bukkit.attribute.Attribute;
 import org.bukkit.entity.LivingEntity;
 import org.bukkit.entity.Player;
@@ -33,7 +32,7 @@ import org.bukkit.inventory.meta.ItemMeta;
 
 import pl.plajer.pinata.Main;
 import pl.plajer.pinata.pinata.Pinata;
-import pl.plajer.pinata.pinataapi.PinataFactory;
+import pl.plajer.pinata.api.PinataFactory;
 import pl.plajer.pinata.utils.PinataUtils;
 import pl.plajer.pinata.utils.Utils;
 import pl.plajerlair.core.utils.XMaterial;
